Type:Addition
Changes: I made the minus and the plus buttons work and the sliders are now independently working, i made a function that updates the colors. I documented a bit in the code, made it more readable and less complicated. I also am not working with the 0.001 value but just whole numbers, it only gets to that value when its for the user to see or when the colors need to be updated. for the rest its just whole numbers(INT)

Bugs: i saw that when i go over the 1 value it gives an NaN error in the js, but i dont know how to fix it right now. Sometimes it goes above 1 like 1.001 or -0.001 and this too i dont know how to fix right now.