Type:Addition
Changes: With the help of chat gpt i have made a formula to calculate the difference of the 6 sliders and give a score. i added colors to the score too, to make the app more alive. i think this was the finale of this color guessing game in javascript. i dont know if there can be more things added to it.